+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for `blinools`
 
+## UNRELEASED
+
+* Introduce new `prune` command
+
 ## 0.1.0
 
 * Initial release